chore: Refactor HR tests
